Conflicts:
	ChangeLog
	SConstruct
	libdrqueue/common.h
	python/setup.py

+++ blockhost.c
@@ -1,5 +1,8 @@
 //
 // Copyright (C) 2001,2002,2003,2004 Jorge Daza Garcia-Blanes
+// Copyright (C) 2010 Andreas Schroeder
+//
+// This file is part of DrQueue
 //
 // This program is free software; you can redistribute it and/or modify
 // it under the terms of the GNU General Public License as published by
@@ -15,8 +18,6 @@
 // along with this program; if not, write to the Free Software
 // Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307
 // USA
-//
-// $Id$
 //
 
 #include <stdio.h>
@@ -39,8 +40,8 @@ void usage (void);
 
 int main (int argc,char *argv[]) {
   int opt;
-  uint32_t ijob = -1;
-  uint32_t icomp = -1;
+  int ijob = -1;
+  int icomp = -1;
   int action = ACTION_NONE;
   struct blocked_host *bh;
   uint16_t nblocked;
@@ -86,21 +87,21 @@ int main (int argc,char *argv[]) {
 
   switch (action) {
   case ACTION_ADD:
-    if (!request_job_add_blocked_host (ijob,icomp,CLIENT)) {
+    if (!request_job_add_blocked_host ((uint32_t)ijob,(uint32_t)icomp,CLIENT)) {
       fprintf (stderr,"ERROR: While trying to add host to block list: %s\n",drerrno_str());
       exit (1);
     }
     printf ("Host blocked successfully\n");
     break;
   case ACTION_DEL:
-    if (!request_job_delete_blocked_host (ijob,icomp,CLIENT)) {
+    if (!request_job_delete_blocked_host ((uint32_t)ijob,(uint32_t)icomp,CLIENT)) {
       fprintf (stderr,"ERROR: While trying to delete host from block list: %s\n",drerrno_str());
       exit (1);
     }
     printf ("Host unblocked successfully\n");
     break;
   case ACTION_LIST:
-    request_job_list_blocked_host (ijob,&bh,&nblocked,CLIENT);
+    request_job_list_blocked_host ((uint32_t)ijob,&bh,&nblocked,CLIENT);
     for (i=0;i<nblocked;i++) {
       printf ("%i\t%s\n",i,bh[i].name);
     }
